Player Actions Functionality
Added betting, calling, spotting, bidding aces server calls.

We don't have turn tracking as far as I can tell at the moment, so the functions for call, spot, and bidAces need to be finished with the proper values of the previous opponent's bid data.

We also need to hide the player actions when it's an opponent's turn. This will come with turn tracking as the above.